requirements.txt
fastapi==0.111.0
pydantic-settings==2.3.3
typing==3.7.4.3
cryptography==42.0.8
pyjwt==2.8.0
pandas==2.2.0
pyxlsb==1.0.10
openpyxl==3.1.2
python-multipart==0.0.9
PyPDF2==3.0.1
uvicorn==0.27.1
python-dotenv==1.0.1
office365-rest-python-client==2.5.0
aiohttp==3.9.3
